# [OTE-529] feat(OE): Deprecate in-mem Untriggered Conditional Orders 