--- a/protocol/x/clob/keeper/final_settlement.go
+++ b/protocol/x/clob/keeper/final_settlement.go
@@ -14,9 +14,6 @@ import (
 func (k Keeper) mustTransitionToFinalSettlement(ctx sdk.Context, clobPairId types.ClobPairId) {
 	// Forcefully cancel all stateful orders from state for this clob pair.
 	k.mustCancelStatefulOrdersForFinalSettlement(ctx, clobPairId)
-
-	// Delete untriggered conditional orders for this clob pair from memory.
-	delete(k.UntriggeredConditionalOrders, clobPairId)
 }
 
 // mustCancelStatefulOrdersForFinalSettlement forcefully cancels all stateful orders