--- a/protocol/x/clob/keeper/untriggered_conditional_orders.go
+++ b/protocol/x/clob/keeper/untriggered_conditional_orders.go
@@ -52,43 +52,6 @@ func (untriggeredOrders *UntriggeredConditionalOrders) IsEmpty() bool {
 		len(untriggeredOrders.OrdersToTriggerWhenOraclePriceGTETriggerPrice) == 0
 }
 
-// AddUntriggeredConditionalOrders takes in a list of newly-placed conditional order ids and adds them
-// to the in-memory UntriggeredConditionalOrders struct, filtering out orders that have been cancelled
-// or expired in the last block. This function is used in EndBlocker and on application startup.
-func (k Keeper) AddUntriggeredConditionalOrders(
-	ctx sdk.Context,
-	placedConditionalOrderIds []types.OrderId,
-	placedStatefulCancellationOrderIds map[types.OrderId]struct{},
-	expiredStatefulOrderIdsSet map[types.OrderId]struct{},
-) {
-	for _, orderId := range placedConditionalOrderIds {
-		_, isCancelled := placedStatefulCancellationOrderIds[orderId]
-		_, isExpired := expiredStatefulOrderIdsSet[orderId]
-		if isCancelled || isExpired {
-			continue
-		}
-
-		orderPlacement, exists := k.GetUntriggeredConditionalOrderPlacement(ctx, orderId)
-		if !exists {
-			panic(
-				fmt.Sprintf(
-					"AddUntriggeredConditionalOrders: order placement does not exist in state for untriggered "+
-						"conditional order id, OrderId %+v.",
-					orderId,
-				),
-			)
-		}
-
-		clobPairId := types.ClobPairId(orderId.GetClobPairId())
-		untriggeredConditionalOrders, exists := k.UntriggeredConditionalOrders[clobPairId]
-		if !exists {
-			untriggeredConditionalOrders = k.NewUntriggeredConditionalOrders()
-			k.UntriggeredConditionalOrders[clobPairId] = untriggeredConditionalOrders
-		}
-		untriggeredConditionalOrders.AddUntriggeredConditionalOrder(orderPlacement.GetOrder())
-	}
-}
-
 // AddUntriggeredConditionalOrder adds an untriggered conditional order to the UntriggeredConditionalOrders
 // data structure. It will panic if the order is not a conditional order.
 func (untriggeredOrders *UntriggeredConditionalOrders) AddUntriggeredConditionalOrder(order types.Order) {
@@ -119,58 +82,6 @@ func (untriggeredOrders *UntriggeredConditionalOrders) AddUntriggeredConditional
 				untriggeredOrders.OrdersToTriggerWhenOraclePriceLTETriggerPrice,
 				order,
 			)
-		}
-	}
-}
-
-// PruneUntriggeredConditionalOrders takes in lists of expired and cancelled stateful order ids and removes
-// all respective orders from the in-memory `UntriggeredConditionalOrders` data structure. This data structure
-// stores untriggered orders in a map of ClobPairId -> []Order, so we first group orders by ClobPairId and then
-// call `UntriggeredConditionalOrders.RemoveExpiredUntriggeredConditionalOrders` on each ClobPairId.
-func (k Keeper) PruneUntriggeredConditionalOrders(
-	expiredStatefulOrderIds []types.OrderId,
-	cancelledStatefulOrderIds []types.OrderId,
-) {
-	// Merge lists of order ids.
-	orderIdsToPrune := lib.UniqueSliceToSet(expiredStatefulOrderIds)
-	for _, orderId := range cancelledStatefulOrderIds {
-		if _, exists := orderIdsToPrune[orderId]; exists {
-			panic(
-				fmt.Sprintf(
-					"PruneUntriggeredConditionalOrders: duplicate order id %+v in expired and "+
-						"cancelled order lists", orderId,
-				),
-			)
-		}
-		orderIdsToPrune[orderId] = struct{}{}
-	}
-
-	prunableUntriggeredConditionalOrderIdsByClobPair := make(map[types.ClobPairId][]types.OrderId)
-	for orderId := range orderIdsToPrune {
-		// If the order id is conditional, add to prunable list of untriggered order ids.
-		// Triggered conditional orders will be effectively ignored during removal as they are not part of
-		// UntriggeredConditionalOrders anymore. No need to filter out here, we can avoid memstore reads.
-		if orderId.IsConditionalOrder() {
-			clobPairId := types.ClobPairId(orderId.GetClobPairId())
-			if _, exists := prunableUntriggeredConditionalOrderIdsByClobPair[clobPairId]; !exists {
-				prunableUntriggeredConditionalOrderIdsByClobPair[clobPairId] = []types.OrderId{}
-			}
-
-			prunableUntriggeredConditionalOrderIdsByClobPair[clobPairId] = append(
-				prunableUntriggeredConditionalOrderIdsByClobPair[clobPairId],
-				orderId,
-			)
-		}
-	}
-
-	for clobPairId := range prunableUntriggeredConditionalOrderIdsByClobPair {
-		if untriggeredConditionalOrders, exists := k.UntriggeredConditionalOrders[clobPairId]; exists {
-			untriggeredConditionalOrders.RemoveUntriggeredConditionalOrders(
-				prunableUntriggeredConditionalOrderIdsByClobPair[clobPairId],
-			)
-			if untriggeredConditionalOrders.IsEmpty() {
-				delete(k.UntriggeredConditionalOrders, clobPairId)
-			}
 		}
 	}
 }
@@ -264,6 +175,26 @@ func (untriggeredOrders *UntriggeredConditionalOrders) PollTriggeredConditionalO
 	return triggeredOrderIds
 }
 
+// OrganizeUntriggeredConditionalOrdersFromState takes in a list of conditional orders read from
+// state, organize them and return in form of `UntriggeredConditionalOrders` struct.
+func OrganizeUntriggeredConditionalOrdersFromState(
+	conditonalOrdersFromState []types.Order,
+) map[types.ClobPairId]*UntriggeredConditionalOrders {
+	ret := make(map[types.ClobPairId]*UntriggeredConditionalOrders)
+
+	for _, order := range conditonalOrdersFromState {
+		clobPairId := types.ClobPairId(order.GetClobPairId())
+		untriggeredConditionalOrders, exists := ret[clobPairId]
+		if !exists {
+			untriggeredConditionalOrders = NewUntriggeredConditionalOrders()
+			ret[clobPairId] = untriggeredConditionalOrders
+		}
+		untriggeredConditionalOrders.AddUntriggeredConditionalOrder(order)
+	}
+
+	return ret
+}
+
 // MaybeTriggerConditionalOrders queries the prices module for price updates and triggers
 // any conditional orders in `UntriggeredConditionalOrders` that can be triggered. For each triggered
 // order, it takes the stateful order placement stored in Untriggered state and moves it to Triggered state.
@@ -278,15 +209,19 @@ func (k Keeper) MaybeTriggerConditionalOrders(ctx sdk.Context) (allTriggeredOrde
 		time.Now(),
 	)
 
+	clobPairToUntriggeredConditionals := OrganizeUntriggeredConditionalOrdersFromState(
+		k.GetAllUntriggeredConditionalOrders(ctx),
+	)
+
 	// Sort the keys for the untriggered conditional orders struct. We need to trigger
 	// the conditional orders in an ordered way to have deterministic state writes.
-	sortedKeys := lib.GetSortedKeys[types.SortedClobPairId](k.UntriggeredConditionalOrders)
+	sortedKeys := lib.GetSortedKeys[types.SortedClobPairId](clobPairToUntriggeredConditionals)
 
 	allTriggeredOrderIds = make([]types.OrderId, 0)
 	// For all clob pair ids in UntriggeredConditionalOrders, fetch the updated
 	// oracle price and poll out triggered conditional orders.
 	for _, clobPairId := range sortedKeys {
-		untriggered := k.UntriggeredConditionalOrders[clobPairId]
+		untriggered := clobPairToUntriggeredConditionals[clobPairId]
 		clobPair, found := k.GetClobPair(ctx, clobPairId)
 
 		// Error log and skip to next clob pair id if invalid clob pair id found.
@@ -322,9 +257,6 @@ func (k Keeper) MaybeTriggerConditionalOrders(ctx sdk.Context) (allTriggeredOrde
 			triggered = k.TriggerOrdersWithPrice(ctx, untriggered, clampedMaxTradePrice, perpetualId, metrics.MaxTradePrice)
 			allTriggeredOrderIds = append(allTriggeredOrderIds, triggered...)
 		}
-
-		// Set the modified untriggeredConditionalOrders back on the keeper field.
-		k.UntriggeredConditionalOrders[clobPairId] = untriggered
 
 		// Gauge the number of untriggered orders.
 		metrics.SetGaugeWithLabels(