### Changelist
Deprecate `UntriggeredConditionalOrders` object on the ClobKeeper and triggers based on state instead. This is required for OE integration. Since this slightly increases `EndBlocker` latency (see results below), will hold-off on merging until OE is ready to be released. More context [here](https://dydx-team.slack.com/archives/C078HSZUV6E/p1719435132581169) 

### Test Plan
Tested this implementation on mainnet full node ([relevant period](https://app.datadoghq.com/s/dc160ddf0/q4n-ud9-9kc), [details](https://dydx-team.slack.com/archives/C078HSZUV6E/p1719932626448099?thread_ts=1719435132.581169&cid=C078HSZUV6E)): 
- No app hash issue so change is non-breaking
- [p50 UCO trigger latency](https://app.datadoghq.com/dashboard/25s-bg4-rfr?fromUser=true&refresh_mode=paused&tile_focus=1510530326906014&view=spans&from_ts=1719845863680&to_ts=1719932263680&live=false) increased from 0.8ms -> 2.3ms; p99 1.2ms->3.6ms. 

### Author/Reviewer Checklist
- [ ] If this PR has changes that result in a different app state given the same prior state and transaction list, manually add the `state-breaking` label.
- [ ] If the PR has breaking postgres changes to the indexer add the `indexer-postgres-breaking` label.
- [ ] If this PR isn't state-breaking but has changes that modify behavior in `PrepareProposal` or `ProcessProposal`, manually add the label `proposal-breaking`.
- [ ] If this PR is one of many that implement a specific feature, manually label them all `feature:[feature-name]`.
- [ ] If you wish to for mergify-bot to automatically create a PR to backport your change to a release branch, manually add the label `backport/[branch-name]`.
- [ ] Manually add any of the following labels: `refactor`, `chore`, `bug`.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **New Features**
    - Added new conditional order types for various buying and selling scenarios with specific parameters.

- **Bug Fixes**
    - Improved management of conditional orders to enhance system reliability and performance.

- **Refactor**
    - Simplified the logic for managing untriggered conditional orders and removed unnecessary complexity.

- **Tests**
    - Updated and added test cases to validate changes in conditional order handling and ensure robustness of new features.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->